+++ README.adoc
@@ -11,11 +11,8 @@ image:https://img.shields.io/badge/Use JakartaEE-project-brightgreen.svg?style=p
 [[readme-en]]
 == Frapp:ee Documentation
 
-Frapp:ee is a collection of implementations of coff:ee solutions that build on implementations of the JakartaEE specificaiton.
-Its structure is modular, and almost everything can be overridden at the project level. The framework is based on 
-coff:ee and its dependencies, completed with the following:
-
-* Hibernate
+Frapp:ee is a collection of ORM solutions. Its structure is modular, and almost everything can be 
+overridden at the project level. The framework is based on coffee-ee modules.
 
 == Maven central
 Frapp:ee can also be found in the official maven repository as well, 
@@ -41,7 +38,7 @@ it is enough to reference necessary BOM objects in the project which handles eac
 <dependencies>
     <dependency>
         <groupId>hu.icellmobilsoft.frappee</groupId>
-        <artifactId>frappee-hibernate</artifactId>
+        <artifactId>frappee-hibernate-batch</artifactId>
     </dependency>
     ...other frapp:ee modules...
 </dependencies>
@@ -55,11 +52,8 @@ Copyright (C) 2020 i-Cell Mobilsoft Zrt.
 :sectnums:
 == Frapp:ee Dokumentáció
 
-A frapp:ee olyan coff:ee megoldások implementációs gyűjteménye, melyek építenek a JakartaEE specifikáció implementációira.
-A felépítése moduláris, és felülírható projekt szinten szinte minden. A keretrendszer a coff:ee egyes részeire és
-annak függőségeire épít, kiegészítve az alábbiakkal:
-
-* Hibernate
+A frapp:ee ORM megoldások implementációs gyűjteménye. A felépítése moduláris, és felülírható projekt
+szinten szinte minden. A keretrendszer a coffee-se moduljaira épít.
 
 == Maven central
 A frapp:ee megtalálható az official maven repository-ban is,
@@ -85,7 +79,7 @@ a projekten elég behúzni a BOM-ot mely minden almodult lekezel:
 <dependencies>
     <dependency>
         <groupId>hu.icellmobilsoft.frappee</groupId>
-        <artifactId>frappee-hibernate</artifactId>
+        <artifactId>frappee-hibernate-batch</artifactId>
     </dependency>
     ...egyéb frappee:ee modulok...
 </dependencies>
@@ -94,4 +88,3 @@ a projekten elég behúzni a BOM-ot mely minden almodult lekezel:
 Bővebb leírás a http://i-cell-mobilsoft-open-source.github.io/frappee[frapp:ee dokumentáció] oldalon található.
 
 Copyright (C) 2020 i-Cell Mobilsoft Zrt.
-
