+++ docs/en/modules.adoc
@@ -1,3 +1,4 @@
 [#modules]
 
-include::modules/frappee-modules.adoc[leveloffset=+1]
+include::modules/frappee-jpa.adoc[leveloffset=+1]
+include::modules/frappee-hibernate.adoc[leveloffset=+1]
