+++ docs/en/project-description.adoc
@@ -1,7 +1,4 @@
 == Frapp:ee Documentation
 
-Frapp:ee is a collection of implementations of coff:ee solutions that build on implementations of the JakartaEE specificaiton.
-Its structure is modular, and almost everything can be overridden at the project level. The framework is based on 
-coff:ee and its dependencies, completed with the following:
-
-* Hibernate
+Frapp:ee is a collection of ORM solutions. Its structure is modular, and almost everything can be 
+overridden at the project level. The framework is based on coffee-ee modules.
