+++ docs/hu/migration.adoc
@@ -1,2 +1,3 @@
 [#migration]
+
 = Migrációs leírások
