+++ docs/hu/project-description.adoc
@@ -1,7 +1,4 @@
-== Frapp:ee Dokumentáció
+= Frapp:ee Dokumentáció
 
-A frapp:ee olyan coff:ee megoldások implementációs gyűjteménye, melyek építenek a JakartaEE specifikáció implementációira.
-A felépítése moduláris, és felülírható projekt szinten szinte minden. A keretrendszer a coff:ee egyes részeire és
-annak függőségeire épít, kiegészítve az alábbiakkal:
-
-* Hibernate
+A frapp:ee ORM megoldások implementációs gyűjteménye. A felépítése moduláris, és felülírható projekt
+szinten szinte minden. A keretrendszer a coffee-se moduljaira épít.
