+++ docs/hu/samples.adoc
@@ -1,0 +1,5 @@
+[#samples]
+
+= Példák
+
+include::samples/frappee-hibernate-samples.adoc[]
