+++ frappee-bom/frappee-bom-all/pom.xml
@@ -14,8 +14,8 @@
 
     <properties>
         <version.hu.icellmobilsoft.coffee>2.8.0-SNAPSHOT</version.hu.icellmobilsoft.coffee>
-        <version.org.hibernate>6.2.13.Final</version.org.hibernate> <!-- WF
-        30.0.1 -->
+        <version.org.hibernate>6.2.13.Final</version.org.hibernate> <!-- Wildfly 30.0.1 -->
+        <version.org.apache.commons.commons-lang3>3.12.0</version.org.apache.commons.commons-lang3>
     </properties>
 
     <dependencyManagement>
@@ -25,6 +25,13 @@
             <dependency>
                 <groupId>hu.icellmobilsoft.frappee</groupId>
                 <artifactId>frappee-bom-project</artifactId>
+                <version>${project.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+            <dependency>
+                <groupId>hu.icellmobilsoft.frappee</groupId>
+                <artifactId>frappee-bom-jakarta</artifactId>
                 <version>${project.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
@@ -52,6 +59,13 @@
                 <artifactId>hibernate-core</artifactId>
                 <version>${version.org.hibernate}</version>
                 <scope>provided</scope>
+            </dependency>
+
+			<!-- Utils -->
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-lang3</artifactId>
+                <version>${version.org.apache.commons.commons-lang3}</version>
             </dependency>
 
         </dependencies>
