+++ frappee-hibernate/frappee-hibernate-batch-cdi/src/main/resources/META-INF/beans.xml
