+++ frappee-hibernate/frappee-hibernate-batch/pom.xml
@@ -1,0 +1,47 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>hu.icellmobilsoft.frappee</groupId>
+        <artifactId>frappee-hibernate</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>frappee-hibernate-batch</artifactId>
+
+    <dependencies>
+
+		<dependency>
+			<groupId>hu.icellmobilsoft.coffee</groupId>
+			<artifactId>coffee-se-util</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>hu.icellmobilsoft.frappee</groupId>
+			<artifactId>frappee-jpa-batch-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>hu.icellmobilsoft.frappee</groupId>
+			<artifactId>frappee-hibernate-util</artifactId>
+		</dependency>
+		
+        <dependency>
+            <groupId>org.hibernate.orm</groupId>
+            <artifactId>hibernate-core</artifactId>
+        </dependency>
+
+        <!--Test dependecies -->
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-junit-jupiter</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
