+++ frappee-hibernate/frappee-hibernate-batch/src/main/java/hu/icellmobilsoft/frappee/hibernate/batch/types/StringBasicType.java
