+++ frappee-hibernate/frappee-hibernate-batch/src/test/java/hu/icellmobilsoft/frappee/hibernate/batch/provider/BinaryArgumentsProvider.java
