+++ frappee-hibernate/frappee-hibernate-util-cdi/pom.xml
@@ -1,0 +1,33 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>hu.icellmobilsoft.frappee</groupId>
+		<artifactId>frappee-hibernate</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>frappee-hibernate-util-cdi</artifactId>
+
+	<dependencies>
+
+		<dependency>
+			<groupId>hu.icellmobilsoft.frappee</groupId>
+			<artifactId>frappee-hibernate-util</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>jakarta.enterprise</groupId>
+			<artifactId>jakarta.enterprise.cdi-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>jakarta.persistence</groupId>
+			<artifactId>jakarta.persistence-api</artifactId>
+		</dependency>
+
+	</dependencies>
+
+</project>
